+++ src/components/AppBar/ActionButton/ActionButton.props.ts
@@ -1,19 +1,6 @@
 import React from 'react';
 
-type AnchorAttributes = {
-  href: string;
-  /** default is '_self' */
-  target?: '_blank' | '_self' | '_parent' | '_top';
-  /** default is 'no-referrer-when-downgrade' */
-  referrer?:
-    | 'no-referrer'
-    | 'no-referrer-when-downgrade'
-    | 'origin'
-    | 'origin-when-cross-origin'
-    | 'same-origin'
-    | 'strict-origin-when-cross-origin'
-    | 'unsafe-url';
-};
+import type { AnchorAttributes } from '../../Link';
 
 type ActionButtonProps = {
   children: React.ReactNode;
