+++ src/components/Link/index.ts
@@ -1,0 +1,5 @@
+import Link from './Link';
+import type { AnchorAttributes } from './Link.props';
+
+export default Link;
+export type { AnchorAttributes };
